retriever_service/requirements.txt
requests
redis
Pillow
imagehash
opencv-python-headless